requirements.txt
flask==1.0.2
pymongo==3.7.2
python-keycloak==0.16.0
requests==2.21.0
uwsgi==2.0.18
PyJWT==1.7.1
cryptography==2.6.1
pandas==0.24.2
WTForms==2.2.1
PyYAML==5.1.2
flask-nav==0.6
Flask-Table==0.5.0
Flask-WTF==0.14.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
bravado-core==5.13.1
tornado==6.0.3
connexion==2.3.0
decorator==4.4.0

